+++ benchmark/imatmul/imatmul_benchmark_logic.hpp
@@ -1,11 +1,12 @@
 //
-// SPDX-FileCopyrightText: Copyright 2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 //
 
 #pragma once
 
+#include <algorithm>
 #include <cstddef>
 #include <cstdint>
 #include <functional>
@@ -62,6 +63,10 @@ void kai_benchmark_imatmul(
         lhs_size *= kai_get_sme_vector_length_u32();
         rhs_size *= kai_get_sme_vector_length_u32();
         dst_size *= kai_get_sme_vector_length_u32();
+    } else if (test::cpu_has_sve()) {
+        lhs_size *= kai_get_sve_vector_length_u32();
+        rhs_size *= kai_get_sve_vector_length_u32();
+        dst_size *= kai_get_sve_vector_length_u32();
     }
 
     const Buffer lhs(lhs_size);
@@ -71,8 +76,23 @@ void kai_benchmark_imatmul(
     ImatmulRunner imatmul_runner(imatmul_interface, dst_type);
     imatmul_runner.set_mnk_chunked(m, n, k_chunk_count, k_chunk_length);
 
+    // Some kernels accept an indirection buffer in place of LHS (when takes_indirection == true)
+    // -----------------------------------------------------------------------------
+    // Do the following :
+    // 1. Create a dummy float value for each pointer to point to.
+    // 2. Create a vector of sufficient size initialized with the default value of a pointer to the float.
+    // 3. If takes_indirection == false, indirection buffer is not used so size is irrelevant.
+    // 4. Pass to kernel runner in place of LHS.
+    std::vector<float> dummy_buffer(std::max<size_t>(k_chunk_count, 1), 1.0f);
+    const auto m_step = imatmul_interface.takes_indirection ? imatmul_interface.get_m_step() : 1;
+    std::vector<const float*> indirection_buffer(k_chunk_count * kai_roundup(m, m_step), dummy_buffer.data());
+
     for (auto _ : state) {
-        imatmul_runner.run(lhs.data(), rhs.data(), dst.data());
+        if (imatmul_interface.takes_indirection) {
+            imatmul_runner.run((const void*)indirection_buffer.data(), rhs.data(), dst.data());
+        } else {
+            imatmul_runner.run((const void*)lhs.data(), rhs.data(), dst.data());
+        }
     }
 }